// src/models/AcHeadsModel.tsx
export interface AcHeadsModel{
    SocietyId: string;
    AcHeadId: string;
    AcHead: string;
    SubCategoryId: string;
    Sequence: number;
    Nature: string;
    TdscategoryId: string;
    Tdscompany: boolean;
    Pan: string;
    Address: string;
    City: string;
    Pin: number;
    StateId: string;
    CountryCode: string;
    Active: boolean;
    GstNo: string;
    
}